Added std.concurrency.  The cuurrent implementation is still pretty spare, but the basics work.  See the unittest for an example.
"make unittest" doesn't work with this module because the compiler breaks on what it thinks is unreachable code.  I suppose "treat warnings as errors" will have to be turned off for this module unless someone can find a way to convince the compiler that this code is correct.

More additions forthcoming.  I just wanted to get this in now so it didn't have to continue being passed around via email.