Reverted r1557 (and r1513 partially).
r1557 relaxed some unittests, as some platforms had accuracy problem with strtold().  But the strict unittests were requirements for D; so r1557 should be reverted.

With this change, these unittests will fail on some platforms such as OSX and Gentoo.  However, the unittests SHOULD fail because using strtold() on these platforms is a 'bug' -- we should provide an accurate implementation.

Related issues:
3758: Create D impementation of to!(float, string), etc.
4200: "to!real(to!string(real.min_normal))" raises std.conv.ConvError