Re-enabled std.datetime for Windows but disabled its unit tests for Windows.
For some reason, the std.datetime unit tests compile and pass without
problems on Linux and in Wine on Linux, but on real Windows, the
compiler runs out of memory when compiling them. So, for now, all of the
std.datetime unit tests are versioned with testStdDateTime. On Posix,
version=testdStdDateTime is used in std.datetime, so the tests will run,
but that version is not set on Windows, so the unit tests don't run on
Windows. So, if you want to run the std.datetime unit tests on Windows or
under Wine or Linux, then compile with -version=testStdDateTime.

At the moment, all of the tests pass on both Linux and under Wine, and
the normal code does compile on proper, so I'm fairly certain that it's fine
for use on Windows, but for now, I can't enable the tests on Windows. My
best guess is that it's some sort of weird compiler bug involving system
calls that leak memory on Windows-proper but not in Wine, but I really
don't know.