Some tweaks to some of the std.datetime unit tests.
Also, the Windows-specific unit tests have now been enabled. So, while
unfortunately, we can't currently have them all enabled on Windows,
since it causes dmd to run out of memory, we can at least get the
Windows-specific stuff tested. Hopefully, the fact that the rest are
running on Linux will be enough for now as far as the other tests go.