std.datetime fixed so that it doesn't need the daylight variable on Posix.
It will still use daylight if it's there, but since FreeBSD doesn't have
it for some reason (in spite of the fact that the daylight variable is
part of Posix), we can't assume that it's there.

Also, now that tzname is in druntime, it has been removed from
std.datetime.

This should fix bug# 5616.