Changes to RedBlackTree per suggestions on pull request #10 of D-Programming-Language/Phobos.
I think that I took all of the suggestions into account, so it should be
improved per those suggestions. I also beefed up the unit tests a bit in
order to take different less and allowDuplicate values into account
(previously, it was only testing the defaults for those template
parameters).