Initial commit for the reworking of std.datetime's unit tests as requested.
There should be no functional changes to the code - only the unit tests.
Also, the lines which have been changed are now mostly within 80
characters, and those that are over are not over by much. Only a portion
of the tests have been updated, but I'm doing it in chunks.