Fix broken Windows unittest build
The build has (once again) been broken by increases in object file size.
DMD simply runs out memory while writing the obj file. Unlike bug 4904, when
the symptom was an optlink crash, this time the crash is inside DMD.
The same solution works: split the list of sources up. Previously, the
first list of sources generated >130K fixups; now the first is 99K fixups,
the second 50K, and the third 65K. This should keep us safe for a while.
Note that within a list of sources, the order matters; it's easy to hit
bug 2962 if the order is wrong.