Add denormalizedException
The denormalizedException is set whenever a denormalized value is
manipulated or created. It is useful to know that denormalized values
are generated because it indicates that precision is low.
Further performance of denormalized number can be up to 100 times slower
even when handled in hardware (see Table 2 in "Quantifying the
Interference Caused by Subnormal Floating-Point Values" by Isaac Dooley
and Laxmikant Kale).