Add _d_run_main to give applications more control at start
This allows a application to supply its own main function that
gets run before any D code is run, allowing it to setup external
dependancies that might be tricky to do from static contructors.
GDC has been shipping with something similar sing a long time
so its not especially new or original, the code is compatible.

This is especially usefull for Mac OSX where in order to use Obj-C
code you need to setup various bits of things like a memory pool.

Signed-off-by: Jakob Bornecrantz <wallbraker@gmail.com>