Removed most of the static assert(0, "Unsupported/Unknown OS"); lines.
Now there's only one for the whole file. Upon implementing std.file for
a new OS, for some of the functions, the unit tests will fail to compile
and show that those functions still need to be written. Unfortunately,
not all will be caught that way (since some functions lack unit tests
and some of the unit tests are OS-specific), but it'll cut down on the
extra stuff in the file which may never matter anyway, since we may
never add on OS which is neither Windows nor Posix.