Adjusted std.ctype.hexDigits.
I'm not sure exactly what we want to do with hexDigits. I had originally
added the lowercase a-f to it on the theory that they're valid digits in
hex, but there's code in Phobos which uses std.string.hexdigits.length,
so I adjusted it so that hexDigits is the same as the old hexdigits but
added fullHexDigits for the version with the lowercase letters. Perhaps
folks will have something constructive to say about that in review, but
that's what I'm doing for now.