Removed std.encodingbase.
I was _sure_ that I tested std.encodingbase on Linux, but it's
segfaulting, and after I had added it, I determined that it wasn't
actually needed (simply removing the unused imports for std.encoding in
std.string and std.array fixed the circular dependency with
std.datetime), much as it theoretically would be less likely to cause
problems later to have std.encodingbase. So, for the moment, I'm
removing it. We need to revisit std.encoding's existance anyway.