Removed alias which the spellchecker will catch.
Base on discussions on the Announce group with regards to deprecation,
it sounds like it would be reasonable to skip the deprecation step with
simple function renaming that the compiler will catch and tell you the
corrected name with its spellchecker. So, I'm removing the alias for
isSymLink and just fixing the name straight up. It's unlikely that all
that many people are using it yet anyway, since it hasn't been around
for very long.