Change options that are valid for getopt
There are some inconsistencies/annoyance of options supported by getopt.
Namely -t v was rejected but is now accepted as it is a very common way
to specify an option.
-timeout=v is now rejected as all other -timeout options already are.
This makes sense because -timeout is neither a proper short option (one
dash and one character) nor a long option (two dashes and more than one
character). --t v and --t=v are now rejected for the same reason.
In sum only the short options -t v and -t=v and the long options
--timeout v and --timeout=v are supported.
Additionally specifying a short options with the syntax -tv is now
only supported for v of numeric type. Because this syntax is convenient
but causes only no confusion for numeric values (e.g. -html5 should not
be confused with --html5).