Revert "Removed new, unnecessary overload of countUntil."
This reverts commit b1765dc0cb5f2e185648935448a2b215d07ed2c2.

Okay. So, I misread what count does and reverted my previous changes
prematurely. So I'm putting them back. count counts the number of
elements in the entire range which match the predicate, not just until
the predicate fails. So, it does _not_ do what the new overload of
countUntil does.