std.socket: Wrap C structs for TimeVal and Linger
Instead of trying to mimic the internal layout of the C timeval and linger
structures (and risking silent breakages due to differences in types,
field order and alignment issues), properly wrap them and provide proxy
properties for accessing their fields. Unfortunately I couldn't find a
neat way to do this neatly while preserving documentation, the deprecation
of the C field names and the new aliases. If Issue 6619 is implemented,
the implementation could be simplified greatly.

This commit also renames std.socket.timeval to TimeVal and marks the old
name as scheduled for deprecation.