Size optimisation of std.datetime (saves 12KB)
Large switch statements have been replaced by data-driven array lookups to reduce the amount of code generated and also improve performance. Using DMD 2.056 on OSX x86 this cuts 12KB off of a Hello World program.

The motivation for this change is to try and reduce the size of small D programs as their large size is a common (and valid) complaint amongst newcomers.