Changed deprecation not on std.regexp.
One of the overloads of std.file.listDir still uses std.regexp and
doesn't have a replacement yet. It's scheduled for deprecation with no
date (since it has no proper replacement yet). So, we can't remove
std.regexp in March as scheduled. We'll remove it after the std.file
issue has been resolved.