std.path: Note all memory allocations in docs
Most functions represent simple and cheap string-slicing operations.  In
some cases, however, new memory must be allocated, and this increases
the cost of the operation by orders of magnitude.  As some of these
cases may not be very obvious, I think it is best they are explicitly
mentioned in the documentation.