Some tweaks to unsigned.
unsigned was severely lacking in unit tests. This remedies that. It also
makes its return types more consistent. Previously, in the case of
signed values, the return type was their Unqualed, unsigned counterpart,
whereas in the case of unsigned values, the return type was exactly the
same as the type passed in. Now, it's the Unqualed, unsigned type in all
cases.