Revert "Deprecate crc32 and move it into std.hash.crc32."
This reverts commit d6c2ef0fb6b9ae3b9c5a7dd702261900fee887d7.

These changes are valuable, but after further discussion, we've decided
to review how std.hash looks in general and redesign it, so getting
people to switch from crc32 to std.hash.crc32 will only mean that
they'll have to change their code _again_ once we've sorted out the
std.hash redesign, and we don't want that. So, I'm reverting these
changes for now, and they can become part of the std.hash redesign.