Fix unnecessary template bloat in enforce.
The version of enforce which takes the file and line number as template
arguments has been scheduled for deprecation and has been replaced with
one which takes them as function arguments. The only code that this will
effect is code which explicitly passes the file or line number to
enforce.