bitfields updates (further)
issues dealt with:
4935 - std.bitmanip: bitfields!() template with trailing unnamed field
does not work
8474 - bitfields doesn't work with 32 bit fields
6686 - bitmanip bitfields are broken at 64 bits
5942 - bitfields overwritten erroneously:
5520 - bitfieldsOn
4425 - More bells & whistles for bitfields

Added static check to revert a bitfield to normal type if it is the only
one (and at full size). 32bit as first field no longer errors. Bitfields
should be separated and the no longer overwrite the whole thing (masking
fixed).

bitfields are now pure, safe and nothrow.

bitfieldsOn added (not pure).