Made stride, strideBack, and decode work with ranges of code units.
They're needed for something like a lexer which needs to operate on
arbitrary ranges of code units rather than code points.

I also added versions of stride and strideBack which don't require an
index (and operate on the first and last index of the range
respectively) as well as decodeFront for decoding the first code point
in the range (unlike stride and strideBack, it needs its own name,
because its signature is more or less identical to decode's signature -
the only difference is that its index is out instead of ref).