bitfields - 4525 more bells and whistles
bitfields now allow const as part of it's signature for a bitfield;
Namely the setter will be absent.

for bitfieldsOn template, it should (when referencing a variable)
determine if it's a acceptible data type, and if it's big enough for the
fields. Since stringof doesn't reference it's whole path, a text 'raw'
version is supplied as well for bitfieldsOn (but doesn't do the checks).

Constraints added to bitfieldsOn to only accept built in types, non
float.