Rename std.typetuple to std.generictuple
Git note:
File renaming and creating a new file with same name must be a separated into different commits to allow rebasing this branch and viewing correct diff because git doesn’t explicitly track file movement. Note that final diff of a merge commit of this branch will still be inaccurate just like an automatic rebase of another branch to this commit.