std.math: Use intrinsics from core.math.
This allows us to remove some of the last Phobos-specific bits of
compiler magic. It is also useful because other compilers (e.g. LDC)
implement intrinsics in a different way, which made shipping a modified
copy of std.math necessary for them.

This change moves the compiler dependencies to druntime, where they are
expected. We need the functions in druntime anyway so that alternative
libraries can have the same status as Phobos.