Uncomment unwanted overload of Socket.select.
_No_ D function in druntime or Phobos should be taking time values as
naked integral values. I thought that we'd caught them all, but
Socket.select still has such an overload unfortunately. Given that we're
trying not to break anything right now and deprecated still triggers
errors rather than warnings, I simply uncommented that overload for now.
That way, new code won't use it. We may or may not actually deprecate it
later.

I'm not particularly enthused about the TimeVal type existing in
std.socket either, since it's just duplicating core.time.Duration's
functionality, but I'm not going to try and get rid of it right now.
However, I _did_ move select's Duration overload so that it was first
with the idea that that would increase the odds of people using that
one instead. No functionality should have changed in this commit.