Make std.container.make alias-able.
make is particularly useful for handling the construction of different
range types in unit tests, but it works much better if it can be aliased
so that the construction is fully generic. Also, the fact that make has
not been able to construct structs with no arguments is needlessly
restrictive.