cycle overflow fix.
The problem is that if the cycle is iterated upon for more than size_t.max, the index makes a jump and becomes out of sync, given the modulo approach to indexing. This is especially relevant, since cycle is supposed to be used for circular buffers, which can and are iterated upon for potentially MASSIVE amounts of data.

The new approach is to always bind _index between 0 and range.length, so that _index never actually overflows.

This has the added advantage of shifting the modulo cost from front to popFront(), which is more efficient.

Tweaked a few other things.