Ensure adequate code coverage from std.random.RandomSample unittests.
The old unittests were missing coverage of large sections of
RandomSample.skip() because the ratios of sample size to source size
were such that Algorithm A was always being chosen over Algorithm D.

The new unittests fix this and also extend coverage to a number of
other areas previously not addressed.  There are also currently-failing
tests commented out with version(none) that can be used to verify
future bugfixes and improvements.

A few lines in skip() remain uncovered simply because they depend on
runtime conditions that are extremely rare.